Refactored service-worker.ts file into submodules
