--- src/service-worker.ts
@@ -1,240 +1,0 @@
-/// <reference lib="webworker" />
-// This service worker can be customized!
-// See https://developers.google.com/web/tools/workbox/modules
-// for the list of available Workbox modules, or add any other
-// code you'd like.
-// You can also remove this file if you'd prefer not to use a
-// service worker, and the Workbox build step will be skipped.
-
-import { clientsClaim } from 'workbox-core'
-import { ExpirationPlugin } from 'workbox-expiration'
-import { createHandlerBoundToURL, precacheAndRoute } from 'workbox-precaching'
-import { registerRoute } from 'workbox-routing'
-import { StaleWhileRevalidate } from 'workbox-strategies'
-import { StoredMediaFile, IndexedDBService } from 'services/indexedDbService'
-
-declare const self: ServiceWorkerGlobalScope
-
-const db = new IndexedDBService('firstVoicesIndexedDb')
-
-clientsClaim()
-
-// Precache all of the assets generated by your build process.
-// Their URLs are injected into the manifest variable below.
-// This variable must be present somewhere in your service worker file,
-// even if you decide not to use precaching. See https://cra.link/PWA
-precacheAndRoute(self.__WB_MANIFEST)
-
-// Set up App Shell-style routing, so that all navigation requests
-// are fulfilled with your index.html shell. Learn more at
-// https://developers.google.com/web/fundamentals/architecture/app-shell
-const fileExtensionRegexp = /[^/?]+\.[^/]+$/
-registerRoute(
-  // Return false to exempt requests from being fulfilled by index.html.
-  ({ request, url }: { request: Request; url: URL }) => {
-    // If this isn't a navigation, skip.
-    if (request.mode !== 'navigate') {
-      return false
-    }
-
-    // If this is a URL that starts with /_, skip.
-    if (url.pathname.startsWith('/_')) {
-      return false
-    }
-
-    // If this looks like a URL for a resource, because it contains
-    // a file extension, skip.
-    if (fileExtensionRegexp.exec(url.pathname)) {
-      return false
-    }
-
-    // Return true to signal that we want to use the handler.
-    return true
-  },
-  createHandlerBoundToURL('/index.html')
-)
-
-// An example runtime caching route for requests that aren't handled by the
-// precache, in this case same-origin .png requests like those from in public/
-registerRoute(
-  // Add in any other file extensions or routing criteria as needed.
-  ({ url }) => {
-    return url.origin === self.location.origin && url.pathname.endsWith('.png')
-  },
-  // Customize this strategy as needed, e.g., by changing to CacheFirst.
-  new StaleWhileRevalidate({
-    cacheName: 'images',
-    plugins: [
-      // Ensure that once this runtime cache reaches a maximum size the
-      // least-recently used images are removed.
-      new ExpirationPlugin({ maxEntries: 50 }),
-    ],
-  })
-)
-
-// This allows the web app to trigger skipWaiting via
-// registration.waiting.postMessage({type: 'SKIP_WAITING'})
-self.addEventListener('message', (event) => {
-  if (event.data?.type === 'SKIP_WAITING') {
-    self.skipWaiting().then(() => {
-      self.clients.matchAll({ type: 'window' }).then((clients) => {
-        clients.forEach((client) => {
-          if (client.url && 'navigate' in client) {
-            client.navigate(client.url).then((navClient) => navClient?.focus())
-          }
-        })
-      })
-    })
-  }
-})
-
-self.addEventListener('install', (event) => {
-  event.waitUntil(
-    (async function () {
-      // Activate the new service worker immediately without waiting
-      self.skipWaiting()
-
-      self.clients.claim()
-    })()
-  )
-})
-
-// Media Handling
-self.addEventListener('fetch', (event) => {
-  const request = event.request
-  const url = new URL(request.url)
-  const normalizedUrl = normalizeUrl(url.href)
-
-  // Do not handle if its not a media file
-  if (!isMediaFile(normalizedUrl)) return
-
-  event.respondWith(
-    (async () => {
-      // Try IndexedDB first
-      try {
-        const storedFile = await db.getMediaFile(normalizedUrl)
-
-        if (storedFile) {
-          const blob = new Blob([storedFile.blob], { type: storedFile.type })
-
-          return new Response(blob, {
-            status: storedFile.status,
-            statusText: storedFile.statusText,
-            headers: storedFile.headers,
-          })
-        }
-      } catch (err) {
-        console.error('service-worker: error getting media file from DB: ', normalizedUrl, err)
-      }
-
-      // If not present in DB, fetch and save
-      try {
-        const updatedCorsReq = new Request(request.url, {
-          mode: 'cors',
-          method: request.method,
-          redirect: request.redirect,
-          headers: request.headers,
-        })
-        const networkResponse = await fetch(updatedCorsReq)
-
-        // Fetch and cache the full file in background for audio/video if it was a range request
-        const isRangeRequest = request.headers.has('range')
-        if (isRangeRequest) {
-          fetchFullFileForCaching(normalizedUrl).catch((err) => {
-            console.error('service-worker: error fetching full media file: ', normalizedUrl, err)
-          })
-
-          return networkResponse
-        }
-
-        // For non-range media (images), save the response
-        if (networkResponse.ok) {
-          try {
-            const blob = await networkResponse.clone().blob()
-            const arrayBuffer = await blob.arrayBuffer()
-            const headers = extractHeaders(networkResponse)
-
-            const storedFile: StoredMediaFile = {
-              blob: arrayBuffer,
-              type: blob.type,
-              headers,
-              status: networkResponse.status,
-              statusText: networkResponse.statusText,
-              downloadedAt: '',
-              lastAccessedAt: '',
-            }
-
-            await db.addMediaFile(normalizedUrl, storedFile)
-          } catch (err) {
-            console.error('service-worker: error saving media file to DB: ', normalizedUrl, err)
-          }
-        }
-
-        return networkResponse
-      } catch (err) {
-        console.error('service-worker: Fetch error', err)
-        return new Response('Offline', {
-          status: 503,
-          statusText: 'Service Unavailable',
-        })
-      }
-    })()
-  )
-})
-
-function isMediaFile(url: string) {
-  return endsWithOneOf(url, ['.jpeg', '.jpg', '.gif', '.png', '.tiff', '.tif', '.mp3', '.wav'])
-}
-
-function endsWithOneOf(text: string, endings: string[]): boolean {
-  for (const ending of endings) {
-    if (text.toLocaleLowerCase().endsWith(ending)) {
-      return true
-    }
-  }
-  return false
-}
-
-function normalizeUrl(url: string): string {
-  const u = new URL(url)
-  u.search = ''
-  u.hash = ''
-  return u.toString()
-}
-
-function extractHeaders(response: Response): Record<string, string> {
-  const headers: Record<string, string> = {}
-  response.headers.forEach((value, key) => {
-    headers[key] = value
-  })
-  return headers
-}
-
-async function fetchFullFileForCaching(url: string) {
-  // Avoid duplicate fetches
-  if (await db.hasMediaFile(url)) return
-
-  const fullResponse = await fetch(url, {
-    method: 'GET',
-    mode: 'cors',
-    headers: {}, // To request for full file and not partial response
-  })
-
-  if (fullResponse.ok) {
-    const blob = await fullResponse.clone().blob()
-    const arrayBuffer = await blob.arrayBuffer()
-    const headers = extractHeaders(fullResponse)
-
-    const stored: StoredMediaFile = {
-      blob: arrayBuffer,
-      type: blob.type,
-      headers,
-      status: fullResponse.status,
-      statusText: fullResponse.statusText,
-      downloadedAt: new Date().toISOString(),
-      lastAccessedAt: new Date().toISOString(),
-    }
-
-    await db.addMediaFile(url, stored)
-  }
-}
